Add a dozen new sequence formats + converters.
These are based on the AlignIO/SeqIO module in biopython. A generic
seqio/align conversion tool is provided in order to convert between all
the datatypes.

Additionally an XML generator is included in generic_*io.py such that
when new formats/support is added or removed, it can be re-run and new
converters automatically generated.